applied Christian Lynbech patch for icon managment: "Here is a diff that adds what I think is improved icon handling. It provides an icon box, ie. a configurable place where icons are stacked. As part of this change, I have changed the default location of icons from (750,50) to (0,0). If the icon box overflows, icons will be created ayt (0,0). I have also decreased the default height from 60 to 20 to minimize the space waste (if the application supplies an icon, it will of course take precedence).
The patch allso adds different kinds of sorting options. Icons can be
sorted by creation time (I have added a new slot in the `icon' class)
or name or class (or whatever else you can think of).

Unrelated to the icon box thing, the patch also contains a set of
application inspection function (see the end of "misc.lisp"), A
disabling of the expression that removes the "Exit" entry in the root
menu so that you can close Eclipse down and finally a new option,
*close-display-p*, that regulates whether exiting Eclipse also kills
the entire X session (if *close-display-p* is nil, Eclipse will just
exit).

It still needs a few things, such as the possibility of disabling the
icon box and other fill orders than :top-right (ie. filling from the
top right corner of the box).

Let me know what you think."

plus small corrections from me (for fullscreen-mode, but it still cause problems).