Fix incorect handling of the states _net_wm_state_{above, below}.
They are permanent as the other states (see inpout.lisp).

It implies that the (setf window-priority) should take those state in account when modifying the stack order (see misc.lisp)

This have as last implication, that we should work with an external list of window for the window circulation keystrokes. Indeed, if one window has the _net_wm_state_above state then it will always be on top of the others. What means, that will stay stuck on this window during circulation. (see minor signature modification of circulate-window in virtual-screen.lisp, and small hack in gestures.lisp)